g3_fit: Find step_lengths more flexibly #131
Firstly, if it's in the report use that.

If that fails, then collate all the actions together and find it in the
resulting environment. This means we don't have to hard code a step ID,
which is going to change in the near future.

Maybe gadget3 should offer a g3_to_formula() to do this, it's a oft'
repeated chunk of code internally.

References: https://github.com/gadget-framework/gadget3/issues/131